refactor(packaging): utils - UpcApi
