+++ app/src/main/java/com/example/ezvault/view/fragment/AddItemFragment.java
@@ -35,6 +35,7 @@ import androidx.navigation.Navigation;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.ezvault.utils.UpcApi;
 import com.example.ezvault.view.adapter.PhotoAdapter;
 import com.example.ezvault.R;
 import com.example.ezvault.utils.camera.GalleryAction;
@@ -51,7 +52,6 @@ import com.example.ezvault.utils.SerialPredictor;
 import com.example.ezvault.model.Tag;
 import com.example.ezvault.model.User;
 import com.example.ezvault.utils.textwatchers.NonEmptyTextWatcher;
-import com.example.ezvault.upcAPI;
 import com.example.ezvault.utils.FileUtils;
 import com.example.ezvault.utils.TaskUtils;
 import com.example.ezvault.data.UserManager;
@@ -82,7 +82,7 @@ import dagger.hilt.android.AndroidEntryPoint;
 @AndroidEntryPoint
 public class AddItemFragment extends Fragment {
 
-    private final com.example.ezvault.upcAPI upcLookup = new upcAPI();
+    private final UpcApi upcLookup = new UpcApi();
     private Button createButton;
 
     private AutoCompleteTextView itemSerial;
@@ -113,7 +113,7 @@ public class AddItemFragment extends Fragment {
 
     private Calendar calendar;
 
-    private upcAPI upcAPI;
+    private UpcApi upcAPI;
 
 
     public AddItemFragment() {
@@ -136,7 +136,7 @@ public class AddItemFragment extends Fragment {
         galleryAction = new GalleryAction(requireActivity());
         getLifecycle().addObserver(galleryAction);
 
-        upcAPI = new upcAPI();
+        upcAPI = new UpcApi();
 
     }
