+++ app/src/main/java/com/example/ezvault/view/fragment/EditItemDetails.java
@@ -38,6 +38,7 @@ import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.Toast;
 
+import com.example.ezvault.utils.UpcApi;
 import com.example.ezvault.viewmodel.SelectedItemViewModel;
 import com.example.ezvault.view.adapter.PhotoAdapter;
 import com.example.ezvault.R;
@@ -50,7 +51,6 @@ import com.example.ezvault.model.Item;
 import com.example.ezvault.utils.SerialPrediction;
 import com.example.ezvault.utils.SerialPredictor;
 import com.example.ezvault.utils.textwatchers.NonEmptyTextWatcher;
-import com.example.ezvault.upcAPI;
 import com.example.ezvault.utils.FileUtils;
 import com.example.ezvault.utils.TaskUtils;
 import com.example.ezvault.data.UserManager;
@@ -409,7 +409,7 @@ public class EditItemDetails extends Fragment {
             GmsBarcodeScanner scanner = GmsBarcodeScanning.getClient(EditItemDetails.this.getActivity(), options);
             scanner.startScan().addOnSuccessListener(
                     barcode -> {
-                        upcAPI api = new upcAPI();
+                        UpcApi api = new UpcApi();
                         api.upcLookup(barcode.getRawValue(), getView().findViewById(R.id.edit_details_description), getActivity());
                     }
             );
